+++ changelog.md
@@ -6,7 +6,9 @@ Features:
 
 * Add an option `--init-command` to execute SQL after connecting (Thanks: [KITAGAWA Yasutaka]).
 * Use InputMode.REPLACE_SINGLE
+* Add a `\clip` special command to copy queries to the system clipboard.
 * Add a special command `\pipe_once` to pipe output to a subprocess.
+
 
 Bug Fixes:
 ----------